separate last slot searcher
