+++ eth2near/eth2near-block-relay-rs/src/eth2near_relay.rs
@@ -2,12 +2,12 @@ use crate::beacon_rpc_client::BeaconRPCClient;
 use crate::config::Config;
 use crate::eth1_rpc_client::Eth1RPCClient;
 use crate::hand_made_finality_light_client_update::HandMadeFinalityLightClientUpdate;
-use crate::relay_errors::{ExecutionPayloadError, MissSyncCommitteeUpdate, NoBlockForSlotError};
+use crate::relay_errors::{MissSyncCommitteeUpdate};
+use crate::last_slot_searcher::LastSlotSearcher;
 use contract_wrapper::eth_client_contract_trait::EthClientContractTrait;
 use eth_types::eth2::LightClientUpdate;
 use eth_types::{BlockHeader, H256};
 use log::{debug, info, trace, warn};
-use std::cmp::{max, min};
 use std::error::Error;
 use std::vec::Vec;
 
@@ -46,8 +46,8 @@ pub struct Eth2NearRelay {
     network: String,
     light_client_updates_submission_frequency_in_epochs: i64,
     max_blocks_for_finalization: u64,
-    enable_binsearch: bool,
     near_network_name: String,
+    last_slot_searcher: LastSlotSearcher,
 }
 
 impl Eth2NearRelay {
@@ -71,8 +71,8 @@ impl Eth2NearRelay {
             light_client_updates_submission_frequency_in_epochs: config
                 .light_client_updates_submission_frequency_in_epochs,
             max_blocks_for_finalization: config.max_blocks_for_finalization,
-            enable_binsearch,
             near_network_name: config.near_network_id.to_string(),
+            last_slot_searcher: LastSlotSearcher::new(enable_binsearch),
         };
 
         if register_relay {
@@ -91,7 +91,7 @@ impl Eth2NearRelay {
             info!(target: "relay", "== New relay loop ==");
 
             let last_eth2_slot_on_eth_chain: u64 = skip_fail!(self.beacon_rpc_client.get_last_slot_number(),  "Fail to get last slot on Eth").as_u64();
-            let mut last_eth2_slot_on_near: u64 = skip_fail!(self.get_last_slot(last_eth2_slot_on_eth_chain), "Fail to get last slot on NEAR");
+            let mut last_eth2_slot_on_near: u64 = skip_fail!(self.last_slot_searcher.get_last_slot(last_eth2_slot_on_eth_chain, &self.beacon_rpc_client, &self.eth_client_contract), "Fail to get last slot on NEAR");
 
             info!(target: "relay", "Last slot on near = {}; last slot on eth = {}",
                   last_eth2_slot_on_near, last_eth2_slot_on_eth_chain);
@@ -341,655 +341,17 @@ impl Eth2NearRelay {
     }
 }
 
-// Implementation of functions for searching last slot on NEAR contract
-impl Eth2NearRelay {
-    fn get_last_slot(&mut self, last_eth_slot: u64) -> Result<u64, Box<dyn Error>> {
-        debug!(target: "relay", "= Search for last slot on near =");
-
-        let finalized_slot = self.eth_client_contract.get_finalized_beacon_block_slot()?;
-        trace!(target: "relay", "Finalized slot on near={}", finalized_slot);
-
-        let last_submitted_slot = self.eth_client_contract.get_last_submitted_slot();
-        trace!(target: "relay", "Last submitted slot={}", last_submitted_slot);
-
-        let slot = max(finalized_slot, last_submitted_slot);
-        trace!(target: "relay", "Init slot for search as {}", slot);
-
-        return if self.enable_binsearch {
-            self.binary_slot_search(slot, finalized_slot, last_eth_slot)
-        } else {
-            self.linear_slot_search(slot, finalized_slot, last_eth_slot)
-        };
-    }
-
-
-    // Search for the slot before the first unknown slot on NEAR
-    // Assumptions:
-    //     (1) start_slot is known on NEAR
-    //     (2) last_slot is unknown on NEAR
-    // Return error in case of problem with network connection
-    fn binary_slot_search(
-        &self,
-        slot: u64,
-        finalized_slot: u64,
-        last_eth_slot: u64,
-    ) -> Result<u64, Box<dyn Error>> {
-        if slot == finalized_slot {
-            return self.binsearch_slot_forward(slot, last_eth_slot + 1);
-        }
-
-        match self.block_known_on_near(slot) {
-            Ok(true) => self.binsearch_slot_forward(slot, last_eth_slot + 1),
-            Ok(false) => self.binsearch_slot_range(finalized_slot, slot),
-            Err(err) => {
-                match err.downcast_ref::<NoBlockForSlotError>() {
-                    Some(_) => {
-                        let (left_slot, slot_on_near) = self.find_left_non_error_slot(slot + 1, last_eth_slot + 1, 1);
-                        match slot_on_near {
-                            true => self.binsearch_slot_forward(left_slot, last_eth_slot + 1),
-                            false => self.binsearch_slot_range(finalized_slot, slot),
-                        }
-                    },
-                    None => Err(err)
-                }
-            }
-        }
-    }
-
-    // Search for the slot before the first unknown slot on NEAR
-    // Assumptions:
-    // (1) start_slot is known on NEAR
-    // (2) last_slot is unknown on NEAR
-    // Return error in case of problem with network connection
-    fn binsearch_slot_forward(&self, slot: u64, max_slot: u64) -> Result<u64, Box<dyn Error>> {
-        let mut current_step = 1;
-        let mut prev_slot = slot;
-        while slot + current_step < max_slot {
-            match self.block_known_on_near(slot + current_step) {
-                Ok(true) => {
-                    prev_slot = slot + current_step;
-                    current_step = min(current_step * 2, max_slot - slot);
-                }
-                Ok(false) => break,
-                Err(err) => {
-                    match err.downcast_ref::<NoBlockForSlotError>() {
-                        Some(_) => {
-                            let (slot_id, slot_on_near) =
-                                self.find_left_non_error_slot(slot + current_step - 1, prev_slot, -1);
-                            if slot_on_near {
-                                prev_slot = slot_id;
-                                current_step = min(current_step * 2, max_slot - slot);
-                            } else {
-                                current_step = slot_id - slot;
-                                break;
-                            }
-                        },
-                        None => return Err(err)
-                    }
-                }
-            }
-        }
-
-        self.binsearch_slot_range(prev_slot, slot + current_step)
-    }
-
-
-    // Search for the slot before the first unknown slot on NEAR
-    // Assumptions:
-    // (1) start_slot is known on NEAR
-    // (2) last_slot is unknown on NEAR
-    // Return error in case of problem with network connection
-    fn binsearch_slot_range(&self, start_slot: u64, last_slot: u64) -> Result<u64, Box<dyn Error>> {
-        let mut start_slot = start_slot;
-        let mut last_slot = last_slot;
-        while start_slot + 1 < last_slot {
-            let mid_slot = start_slot + (last_slot - start_slot) / 2;
-            match self.block_known_on_near(mid_slot) {
-                Ok(true) => start_slot = mid_slot,
-                Ok(false) => last_slot = mid_slot,
-                Err(err) => {
-                    match err.downcast_ref::<NoBlockForSlotError>() {
-                        Some(_) => {
-                            let (left_slot, is_left_slot_on_near) =
-                                self.find_left_non_error_slot(mid_slot - 1, start_slot, -1);
-                            if is_left_slot_on_near {
-                                start_slot = mid_slot;
-                            } else {
-                                last_slot = left_slot;
-                            }
-                        },
-                        None => return Err(err),
-                    }
-                }
-            }
-        }
-
-        Ok(start_slot)
-    }
-
-    // Returns the last slot known with block known on NEAR
-    // Slot -- expected last known slot
-    // finalized_slot -- last finalized slot on NEAR, assume as known slot
-    // last_eth_slot -- head slot on Eth
-    fn linear_slot_search(
-        &self,
-        slot: u64,
-        finalized_slot: u64,
-        last_eth_slot: u64,
-    ) -> Result<u64, Box<dyn Error>> {
-        if slot == finalized_slot {
-            return Ok(self.linear_search_forward(slot, last_eth_slot));
-        }
-
-        match self.block_known_on_near(slot) {
-            Ok(true) => Ok(self.linear_search_forward(slot, last_eth_slot)),
-            Ok(false) => Ok(self.linear_search_backward(finalized_slot, slot)),
-            Err(err) => {
-                match err.downcast_ref::<NoBlockForSlotError>() {
-                    Some(_) => {
-                        let left_slot = self.linear_search_forward(slot, last_eth_slot);
-                        return if left_slot > slot {
-                            Ok(left_slot)
-                        } else {
-                            Ok(self.linear_search_backward(finalized_slot, slot))
-                        }
-                    }
-                    None => Err(err)
-                }
-            }
-        }
-    }
-
-    // Returns the slot before the first unknown block on NEAR
-    // The search range is [slot .. max_slot)
-    // If there is no unknown block in this range max_slot - 1 will be returned
-    // Assumptions:
-    //     (1) block for slot is submitted to NEAR
-    //     (2) block for max_slot is not submitted to NEAR
-    fn linear_search_forward(&self, slot: u64, max_slot: u64) -> u64 {
-        let mut slot = slot;
-        while slot < max_slot {
-            match self.block_known_on_near(slot + 1) {
-                Ok(true) => slot += 1,
-                Ok(false) => break,
-                Err(_) => slot += 1,
-            }
-        }
-
-        slot
-    }
-
-    // Returns the slot before the first unknown block on NEAR
-    // The search range is [last_slot .. start_slot)
-    // If no such block are found the start_slot will be returned
-    // Assumptions:
-    //     (1) block for start_slot is submitted to NEAR
-    //     (2) block for last_slot + 1 is not submitted to NEAR
-    fn linear_search_backward(&self, start_slot: u64, last_slot: u64) -> u64 {
-        let mut slot = last_slot;
-        let mut last_false_slot = slot + 1;
-
-        while slot > start_slot {
-            match self.block_known_on_near(slot) {
-                Ok(true) => break,
-                Ok(false) => {
-                    last_false_slot = slot;
-                    slot -= 1
-                },
-                Err(_) => slot -= 1,
-            }
-        }
-
-        last_false_slot - 1
-    }
-
-    // Find the leftmost non-empty slot. Search range: [left_slot, right_slot).
-    // Returns pair: (1) slot_id and (2) is this block already known on Eth client on NEAR
-    // Assume that right_slot is non-empty and it's block were submitted to NEAR,
-    // so if non correspondent block is found we return (right_slot, false)
-    fn find_left_non_error_slot(&self, left_slot: u64, right_slot: u64, step: i8) -> (u64, bool) {
-        let mut slot = left_slot;
-        while slot != right_slot {
-            match self.block_known_on_near(slot) {
-                Ok(v) => return (slot, v),
-                Err(_) => {
-                    if step > 0 {
-                        slot += 1;
-                    } else {
-                        slot -= 1;
-                    }
-                },
-            };
-        }
-
-        if step > 0 {
-            (slot, false)
-        } else {
-            (slot, true)
-        }
-    }
-
-    // Check if the block for current slot in Eth2 already were submitted to NEAR
-    // Returns Error if slot doesn't contain any block
-    fn block_known_on_near(&self, slot: u64) -> Result<bool, Box<dyn Error>> {
-        trace!(target: "relay", "Check if block with slot={} on NEAR", slot);
-        match self
-            .beacon_rpc_client
-            .get_beacon_block_body_for_block_id(&format!("{}", slot))
-        {
-            Ok(beacon_block_body) => {
-                let hash: H256 = H256::from(
-                    beacon_block_body
-                        .execution_payload()
-                        .map_err(|_| ExecutionPayloadError)?
-                        .execution_payload
-                        .block_hash
-                        .into_root()
-                        .as_bytes(),
-                );
-
-                if self.eth_client_contract.is_known_block(&hash)? {
-                    trace!(target: "relay", "Block with slot={} was found on NEAR", slot);
-                    Ok(true)
-                } else {
-                    trace!(target: "relay", "Block with slot={} not found on Near", slot);
-                    Ok(false)
-                }
-            }
-            Err(err) => {
-                trace!(target: "relay", "Error \"{}\" in getting beacon block body for slot={}", err, slot);
-                Err(err)?
-            }
-        }
-    }
-}
-
-
 #[cfg(test)]
 mod tests {
     use eth_types::BlockHeader;
-    use near_units::*;
-    use workspaces::prelude::*;
-    use workspaces::{network::Sandbox, Account, Contract, Worker};
-    use tokio::runtime::Runtime;
-    use contract_wrapper::eth_client_contract::EthClientContract;
-    use contract_wrapper::sandbox_contract_wrapper::SandboxContractWrapper;
     use eth_types::eth2::LightClientUpdate;
     use log::LevelFilter;
-    use crate::config::Config;
     use crate::eth2near_relay::Eth2NearRelay;
-    use crate::test_utils;
     use crate::beacon_rpc_client::BeaconRPCClient;
     use crate::hand_made_finality_light_client_update::HandMadeFinalityLightClientUpdate;
-    use crate::init_contract::init_contract;
     use crate::logger::SimpleLogger;
     use crate::relay_errors::NoBlockForSlotError;
-
-    const WASM_FILEPATH: &str = "../../contracts/near/res/eth2_client.wasm";
-
-    fn create_contract() -> (Account, Contract, Worker<Sandbox>) {
-        let rt = Runtime::new().unwrap();
-
-        let worker = rt.block_on(workspaces::sandbox()).unwrap();
-        let wasm = std::fs::read(WASM_FILEPATH).unwrap();
-        let contract = rt.block_on(worker.dev_deploy(&wasm)).unwrap();
-
-        // create accounts
-        let owner = worker.root_account().unwrap();
-        let relay_account = rt.block_on(owner
-            .create_subaccount(&worker, "relay_account")
-            .initial_balance(parse_near!("30 N"))
-            .transact()).unwrap()
-            .into_result().unwrap();
-
-        (relay_account, contract, worker)
-    }
-
-    fn get_config() -> Config {
-        Config {
-            beacon_endpoint: "https://lodestar-kiln.chainsafe.io".to_string(),
-            eth1_endpoint: "https://rpc.kiln.themerge.dev".to_string(),
-            total_submit_headers: 8,
-            near_endpoint: "NaN".to_string(),
-            signer_account_id: "NaN".to_string(),
-            path_to_signer_secret_key: "NaN".to_string(),
-            contract_account_id: "NaN".to_string(),
-            network: "kiln".to_string(),
-            contract_type: "near".to_string(),
-            light_client_updates_submission_frequency_in_epochs: 1,
-            max_blocks_for_finalization: 5000,
-            near_network_id: "testnet".to_string(),
-            dao_contract_account_id: None,
-            output_dir: None
-        }
-    }
-
-    fn get_relay(enable_binsearch: bool, from_file: bool) -> Eth2NearRelay {
-        let (relay_account, contract, worker) = create_contract();
-        let contract_wrapper = Box::new(SandboxContractWrapper::new(relay_account, contract, worker));
-        let mut eth_client_contract = EthClientContract::new(contract_wrapper);
-
-        let config = get_config();
-        match from_file {
-            true => test_utils::init_contract_from_files(&mut eth_client_contract),
-            false => init_contract(&config, &mut eth_client_contract).unwrap(),
-        };
-
-        let eth_client_contract = Box::new(eth_client_contract);
-
-        Eth2NearRelay::init(&config, eth_client_contract, enable_binsearch, true)
-    }
-
-    #[test]
-    fn test_block_known_on_near() {
-        let mut relay = get_relay(false, true);
-
-        //1060486 slot without block
-        let is_block_known = relay.block_known_on_near(1060486);
-        if let Ok(_) = is_block_known {
-            panic!();
-        }
-
-        let is_block_known = relay.block_known_on_near(1099360);
-
-        match is_block_known {
-            Ok(is_block_known) => assert!(!is_block_known),
-            Err(_) => panic!(),
-        }
-
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(finalized_slot + 1).unwrap()], finalized_slot + 1).unwrap();
-
-        let is_block_known = relay.block_known_on_near(finalized_slot + 1);
-        match is_block_known {
-            Ok(is_block_known) => assert!(is_block_known),
-            Err(_) => panic!(),
-        }
-    }
-
-    #[test]
-    fn test_find_left_non_error_slot() {
-        let mut relay = get_relay(false, true);
-
-        let (left_non_empty_slot, is_known_block) = relay.find_left_non_error_slot(1060528, 1060532, 1);
-        assert_eq!(left_non_empty_slot, 1060528);
-        assert_eq!(is_known_block, false);
-
-        let (left_non_empty_slot, is_known_block) = relay.find_left_non_error_slot(1060529, 1060532, 1);
-        assert_eq!(left_non_empty_slot, 1060531);
-        assert_eq!(is_known_block, false);
-
-        let (left_non_empty_slot, is_known_block) = relay.find_left_non_error_slot(1060529, 1060530, 1);
-        assert_eq!(left_non_empty_slot, 1060530);
-        assert_eq!(is_known_block, false);
-
-        let (left_non_empty_slot, is_known_block) = relay.find_left_non_error_slot(1060530, 1060532, 1);
-        assert_eq!(left_non_empty_slot, 1060531);
-        assert_eq!(is_known_block, false);
-
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(finalized_slot + 1).unwrap()], finalized_slot + 1).unwrap();
-
-        let (left_non_empty_slot, is_known_block) = relay.find_left_non_error_slot(finalized_slot + 1, finalized_slot + 2, 1);
-        assert_eq!(left_non_empty_slot, finalized_slot + 1);
-        assert_eq!(is_known_block, true);
-    }
-
-    #[test]
-    fn test_linear_search_backward() {
-        let mut relay = get_relay(false, true);
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(finalized_slot + 1).unwrap(),
-                                                     relay.get_execution_block_by_slot(finalized_slot + 2).unwrap()], finalized_slot + 2).unwrap();
-
-        let last_submitted_block = relay.linear_search_backward(finalized_slot + 1, finalized_slot + 10);
-        assert_eq!(last_submitted_block, finalized_slot + 2);
-
-        let mut slot = finalized_slot + 3;
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099363 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-        relay.eth_client_contract.send_headers(&blocks, 1099363).unwrap();
-
-        let last_submitted_block = relay.linear_search_backward(finalized_slot + 1, finalized_slot + 10);
-        assert_eq!(last_submitted_block, 1099364);
-    }
-
-    #[test]
-    fn test_linear_search_forward() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        // Slots without blocks: [1099434, 1099438]
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.linear_search_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099500);
-
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.linear_search_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099500);
-
-        assert_eq!(last_block_on_near, 1099364);
-    }
-
-    #[test]
-    fn test_linear_slot_search() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099363 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-        relay.eth_client_contract.send_headers(&blocks, 1099363).unwrap();
-
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-
-        let last_block_on_near = relay.linear_slot_search(1099363, finalized_slot, 1099500).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.linear_slot_search(1099364, finalized_slot, 1099500).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.linear_slot_search(1099361, finalized_slot, 1099500).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.linear_slot_search(1099368, finalized_slot, 1099500).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_error_on_connection_problem() {
-        let mut relay = get_relay(false, true);
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(finalized_slot + 1).unwrap(),
-                                                     relay.get_execution_block_by_slot(finalized_slot + 2).unwrap()], finalized_slot + 2).unwrap();
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        relay.linear_slot_search(finalized_slot + 1, finalized_slot, 1099500).unwrap();
-    }
-
-    #[test]
-    fn test_binsearch_slot_range() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.binsearch_slot_range(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.binsearch_slot_range(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.binsearch_slot_range(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099364).unwrap();
-        assert_eq!(last_block_on_near, 1099363);
-
-        let last_block_on_near = relay.binsearch_slot_range(1099364, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        if let Ok(_) = relay.binsearch_slot_range(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370) {
-            panic!("binarysearch returns result in unworking network");
-        }
-    }
-
-    #[test]
-    fn test_binsearch_slot_forward() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.binsearch_slot_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.binsearch_slot_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.binsearch_slot_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099364).unwrap();
-        assert_eq!(last_block_on_near, 1099363);
-
-        let last_block_on_near = relay.binsearch_slot_forward(1099364, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        if let Ok(_) = relay.binsearch_slot_forward(relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap() + 1, 1099370) {
-            panic!("binarysearch returns result in unworking network");
-        }
-    }
-
-    #[test]
-    fn test_binsearch_slot_search() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        let finalized_slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.binary_slot_search(finalized_slot + 1, finalized_slot, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.binary_slot_search(finalized_slot + 1, finalized_slot, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.binary_slot_search(finalized_slot + 1, finalized_slot, 1099364).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        let last_block_on_near = relay.binary_slot_search(finalized_slot + 1, finalized_slot, 1099363).unwrap();
-        assert_eq!(last_block_on_near, 1099363);
-
-        let last_block_on_near = relay.binary_slot_search(1099364, finalized_slot, 1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099363);
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        if let Ok(_) = relay.binary_slot_search(finalized_slot + 1, finalized_slot, 1099370) {
-            panic!("binarysearch returns result in unworking network");
-        }
-    }
-
-    #[test]
-    fn test_get_last_slot_binsearch() {
-        let mut relay = get_relay(true, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.get_last_slot(1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.get_last_slot(1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        if let Ok(_) = relay.get_last_slot(1099370) {
-            panic!("binarysearch returns result in unworking network");
-        }
-    }
-
-
-    #[test]
-    fn test_get_last_slot_linearsearch() {
-        let mut relay = get_relay(false, true);
-        let mut slot = relay.eth_client_contract.get_finalized_beacon_block_slot().unwrap();
-        slot += 1;
-
-        let mut blocks: Vec<BlockHeader> = vec![];
-        while slot <= 1099362 {
-            if let Ok(block) = relay.get_execution_block_by_slot(slot) {
-                blocks.push(block)
-            }
-            slot += 1;
-        }
-
-        relay.eth_client_contract.send_headers(&blocks, 1099362).unwrap();
-        let last_block_on_near = relay.get_last_slot(1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099362);
-
-        relay.eth_client_contract.send_headers(&vec![relay.get_execution_block_by_slot(1099363).unwrap()], 1099363).unwrap();
-        let last_block_on_near = relay.get_last_slot(1099370).unwrap();
-        assert_eq!(last_block_on_near, 1099364);
-
-        relay.beacon_rpc_client = BeaconRPCClient::new("http://httpstat.us/504/");
-        if let Ok(_) = relay.get_last_slot(1099370) {
-            panic!("binarysearch returns result in unworking network");
-        }
-    }
+    use crate::test_utils::get_relay;
 
     #[test]
     #[ignore]
@@ -1131,7 +493,7 @@ mod tests {
 
                 assert!(finality_update_slot > finalized_slot);
             },
-            Err(_) => {panic!("Error on get light client update");}
+            Err(_) => { panic!("Error on get light client update"); }
         }
     }
 }
