+++ eth2near/eth2near-block-relay-rs/src/lib.rs
@@ -9,6 +9,7 @@ pub mod hand_made_finality_light_client_update;
 pub mod init_contract;
 pub mod logger;
 pub mod relay_errors;
+pub mod last_slot_searcher;
 
 #[cfg(test)]
 pub mod test_utils;
